pyproject.toml
[project]
name = "infernet-monorepo"
version = "0.1.0-1"
description = "Add your description here"
authors = [
    { name = "ritual", email = "hello@ritual.net" }
]
dependencies = [
    "twine>=5.0.0,<6.0.0",
    "pre-commit>=3.7.0",
    "mkdocs-material>=9.5.19,<10.0.0",
    "mkdocstrings>=0.25.2<0.26",
    "mkdocstrings-python>=1.10.8,<2.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "types-pyyaml>=6.0.12.20240311,<7.0.0",
    "dominate>=2.9.1",
    "python-dotenv>=1.0.1,<2.0.0"
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
exclude = ['**/venv', '**/.venv']

[tool.rye]
managed = true
virtual = true
dev-dependencies = []

[tool.rye.workspace]
members = ["libraries/*"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/infernet_monorepo"]

[tool.ruff]
line-length = 89

[tool.isort]
profile = "black"
skip = ["libraries/ritual_pyarweave", ".venv", "venv", ".gitignore"]


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_file = "pytest_logs.txt"
log_file_level = "INFO"
log_file_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "-vvv"
